// frontend/types/api/adMentorLogList.ts
export interface AdMentorLogList {
  userId: number;
  nickname: string;
  role: string;
  roleDescription: string;
  regionId: number;
  noticeCount: number;
  postCount: number;
  commentCount: number;
  approvedCount: number;
  ignoredCount: number;
  invalidCount: number;
  pendingCount: number;
  rejectedCount: number;
  requestedCount: number;
}